Feat: Automatically hide row indices in `st.dataframe` when row selection is active (#12448)
## Describe your changes

Implements the suggestion that range indices in `st.dataframe` should be
hidden automatically when row selections are active and hide_index is
not set or is not false.

Allows users to enable the index column via `hide_index=False`.

Image render link provided in the original issue:
https://github.com/streamlit/streamlit/issues/12237#issuecomment-3192826364
<!-- If it's a visual change, please include a screenshot or video! -->

```python
import streamlit as st
import pandas as pd


df = pd.DataFrame({
  'first column': [1, 2, 3, 4],
  'second column': [10, 20, 30, 40],
  'third column': [10, 20, 30, 40],
  'fourth column': [10, 20, 30, 40],
})

st.dataframe(
  df,
  on_select="rerun",
  # hide_index=False, # can set hide_index=False to re-enable the range index
  selection_mode="single-row" # works on either single-row or multi-row options
)

```

## GitHub Issue Link (if applicable)
Implementation of the suggestion listed here: #12237 

## Testing Plan

- Unit Tests (JS and/or Python)
- E2E Tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <Lukas.Masuch@gmail.com>
Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>