Add `type` argument to `st.popover` to match `st.button` (#12598)
## Describe your changes

The `st.popover` command currently inserts a button in your app which
shows a popover when clicked. However, unlike a regular `st.button`,
that button cannot be styled as primary/secondary/tertiary.

This PR fixes that. This adds `type="primary"|"secondary"|"tertiary"` to
`st.popover`.

## GitHub Issue Link (if applicable)

None

## Testing Plan

- ~~Explanation of why no additional tests are needed~~
- [x] Unit Tests (JS and/or Python)
- [x] E2E Tests
- ~~Any manual testing needed?~~

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: graphite-app[bot] <96075541+graphite-app[bot]@users.noreply.github.com>