Use `key` as main identity for `st.segmented_control` widget (#12711)
## Describe your changes

If a custom `key` is passed to the `st.segmented_control` widget it will
be used as the main identity. This allows for dynamically changing other
parameters in between reruns without causing the widgets to be recreated
in the frontend.

## GitHub Issue Link (if applicable)

- Related to https://github.com/streamlit/streamlit/issues/11277

## Testing Plan

- Added unit and e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.