Create RenderNodeVisitor (#12814)
## Describe your changes

Now, we introduce a new visitor for rendering the block react node. We
made this separate from the render-tree folder because we did not want
to make React a dependency. This is great because we can break out the
render-tree to a separate package someday.

## Testing Plan

- JS Unit Tests for the RenderNode Visitor
- E2E tests should ensure the rendering logic remains correct

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.