[refactor][TextAlignment] Adds `get_heading` helper for tests and updates `st_header_test` to use this helper replacing index access. (#13033)
## Describe your changes

Opportunistically replacing some index access tests with the newer
pattern of using a helper `get_header` that using the text in the
heading element to find the correct element.

<!-- If it's a visual change, please include a screenshot or video! -->

## GitHub Issue Link (if applicable)

## Testing Plan

Test only.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.